backend/requirements.txt
langchain
langchain-community
langchain-core
unstructured
langchain-text-splitters
faiss-cpu
transformers
Flask
python-dotenv
flask-cors
itsdangerous==2.0.1
sentence-transformers
chromadb
google-generativeai
Flask-SQLAlchemy
google-auth
flask-jwt-extended
Flask-Migrate